Modest improvemens to Student, Teacher; Course can now return constraints
